Fix 'date' -> 'published', fix dates, and don't index hidden blogs.
The correct metadata key is 'published' not 'date',
and valid dates must contain the seconds part (just copy from the old
files please).  Blog posts which are hidden (file contains a leading
dot) will not be listed in the index, but are still accesible through
a direct link (/blog/.{the-blog-file-name}).